% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MilinOccu.R
\name{milinbootstrap}
\alias{milinbootstrap}
\title{Parametric Bootstrap for milinoccu}
\usage{
milinbootstrap(n = 100, Y, formattedpar, Designs_occu, Designs_det, ...)
}
\arguments{
\item{n}{Bootstrap sample size}

\item{Y}{The original data, used to determine missing}

\item{formattedpar}{The formatted parameter estimation}

\item{Designs_occu}{Design list for occupancy, see milinoccu}

\item{Designs_det}{Design list for detection, see milinoccu}

\item{...}{Extra arguments send to optim}
}
\value{
A list, each element is an output list of milinoccu
}
\description{
This function run n paramatric bootstrap to help construct confidence intervals
}
