\name{MiLinOccu-package}
\alias{MiLinOccu-package}
\alias{MiLinOccu}
\docType{package}
\title{
\packageTitle{MiLinOccu}
}
\description{
\packageDescription{MiLinOccu}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MiLinOccu}
\packageIndices{MiLinOccu}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MiLinOccu}

Maintainer: \packageMaintainer{MiLinOccu}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
