\name{MiLinOccu-package}
\alias{MiLinOccu-package}
\alias{MiLinOccu}
\docType{package}
\title{
\packageTitle{MiLinOccu}
}
\description{
\packageDescription{MiLinOccu}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MiLinOccu}
\packageIndices{MiLinOccu}

}
\author{
\packageAuthor{MiLinOccu}

Maintainer: \packageMaintainer{MiLinOccu}
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
# not run
}
